Mainly added multithreading
Which means each function definition in the way will be dispatched to
a seperate thread, you need to specify MAXTHREADS for the amount of
threads to be used. (and probably SILENT=1 too so it won't bottleneck)

Also functions that fail to compile will be skipped (obviously I mean
backend failure - if you have unmatched string it'll still bail out).